# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Twilight Core Explorer API",
    "description": "Read-only public API for the Twilight Core explorer (Phase 9a foundation).",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/api/v1/accounts": {
      "get": {
        "summary": "List accounts (address order)",
        "tags": [
          "accounts"
        ],
        "parameters": [
          {
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "accountKind",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "page"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "address",
                          "accountKind",
                          "firstSeenHeight",
                          "lastSeenHeight",
                          "txCount"
                        ],
                        "properties": {
                          "address": {
                            "type": "string"
                          },
                          "accountKind": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "firstSeenHeight": {
                            "anyOf": [
                              {
                                "pattern": "^\\d+$",
                                "description": "Decimal height/id as string",
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "lastSeenHeight": {
                            "anyOf": [
                              {
                                "pattern": "^\\d+$",
                                "description": "Decimal height/id as string",
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "txCount": {
                            "type": "integer"
                          }
                        },
                        "title": "AccountListItem"
                      }
                    },
                    "page": {
                      "type": "object",
                      "required": [
                        "limit",
                        "nextCursor"
                      ],
                      "properties": {
                        "limit": {
                          "minimum": 1,
                          "type": "integer"
                        },
                        "nextCursor": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{address}": {
      "get": {
        "summary": "Get an account by address",
        "tags": [
          "accounts"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "raw"
              ]
            },
            "in": "query",
            "name": "include",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "address",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "object",
                      "required": [
                        "address",
                        "accountKind",
                        "firstSeenHeight",
                        "lastSeenHeight",
                        "txCount"
                      ],
                      "properties": {
                        "address": {
                          "type": "string"
                        },
                        "accountKind": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "firstSeenHeight": {
                          "anyOf": [
                            {
                              "pattern": "^\\d+$",
                              "description": "Decimal height/id as string",
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "lastSeenHeight": {
                          "anyOf": [
                            {
                              "pattern": "^\\d+$",
                              "description": "Decimal height/id as string",
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "txCount": {
                          "type": "integer"
                        },
                        "raw": {}
                      },
                      "title": "AccountDetail"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}